code reorganisation for better api installation and unit test
