+++ setup.py
@@ -1,12 +1,14 @@
 from setuptools import setup
-
+import api_rte
 setup(
     name='api_rte',
-    version=1.1,
+    version=api_rte.__version__,
     description='Library RTE API',
     author='Vincent Reinbold',
     author_email='vincent.reinbold@geeps.centralesupelec.fr',
     license='GNU GENERAL PUBLIC LICENSE v3',
-    packages=[],
-    classifiers=["Programming Language :: Python :: 3.7"], install_requires=['requests', 'urllib3', 'pandas']
+    packages=['api_rte'],
+    classifiers=["Programming Language :: Python :: 3.7"],
+    install_requires=['requests', 'urllib3', 'pandas'],
+    test_suite='tests',
 )
